+++ test/recording_studio_wordpress_plugin_template_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.4.14", ::RecordingStudioWordpressPluginTemplate::VERSION
+    assert_equal "0.4.15", ::RecordingStudioWordpressPluginTemplate::VERSION
   end
 
   def test_engine_exists
@@ -86,12 +86,15 @@ class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_accessible", tag: "v0.9.1"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_root_switchable", tag: "v0.5.0"'
     assert_includes gemfile, 'github: "bowerbird-app/flatpack", tag: "v0.1.177"'
-    assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_api", tag: "v0.5.5"'
+    assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_api", tag: "v0.5.6"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_Embeddable", tag: "v0.2.1"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_admin", tag: "v2.0.2"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_publishable", tag: "v0.2.0"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_attachable", tag: "v0.5.1"'
-    assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_Oauth", tag: "v0.2.0"'
+    assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_Oauth", tag: "v0.2.2"'
+    lockfile = File.read(File.expand_path("dummy/Gemfile.lock", __dir__))
+    assert_includes lockfile, "recording_studio_oauth (0.2.2)"
+    assert_includes lockfile, "recording_studio_api (0.5.6)"
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_site_settings", tag: "v0.1.0"'
     assert_includes gemfile, 'github: "bowerbird-app/RecordingStudio_users", tag: "v0.11.0"'
     refute_includes gemfile, "recording_studio/v3.0.0"
@@ -280,6 +283,8 @@ class RecordingStudioWordpressPluginTemplateTest < Minitest::Test
     assert_includes readme, "v4.2.0"
     assert_includes readme, "v0.1.177"
     assert_includes readme, "v0.9.1"
+    assert_includes readme, "dummy GitHub tag `v0.2.2`"
+    assert_includes readme, "public clients can exchange codes on the named token URL"
     refute_includes readme, "Internal template"
     refute_includes readme, "internal template"
     refute_includes readme, "v0.1.133"